servicebus01/func/requirements.txt
azure-monitor-opentelemetry 
azure-functions
azure-identity